fix(options): Fix possible memory leak for boost.useWorker
- Release URL object after worker execution is done
- Reuse Blob object for same function

Ref #2778
