+++ src/ChartInternal/data/convert.ts
@@ -2,7 +2,8 @@
  * Copyright (c) 2017 ~ present NAVER Corp.
  * billboard.js project is licensed under the MIT license
  */
-import {runWorker, isUndefined, isDefined, isObject, isValue, notEmpty, isArray} from "../../module/util";
+import {isUndefined, isDefined, isObject, isValue, notEmpty, isArray} from "../../module/util";
+import {runWorker} from "../../module/worker";
 import {columns, json, rows, url} from "./convert.helper";
 import type {IData} from "../data/IData";
