+++ src/module/util.ts
@@ -53,7 +53,6 @@ export {
 	notEmpty,
 	parseDate,
 	runUntil,
-	runWorker,
 	sanitise,
 	setTextValue,
 	sortValue,
@@ -796,78 +795,6 @@ function convertInputType(mouse: boolean, touch: boolean): "mouse" | "touch" | n
 
 	// fallback to 'mouse' if no input type is detected.
 	return (hasMouse && "mouse") || (hasTouch && "touch") || "mouse";
-}
-
-/**
- * Create and run on Web Worker
- * @param {boolean} useWorker Use Web Worker
- * @param {Function} fn Function to be executed in worker
- * @param {Function} callback Callback function to receive result from worker
- * @param {Array} depsFn Dependency functions to run given function(fn).
- * @returns {object}
- * @example
- * 	const worker = runWorker(function(arg) {
- *		  // do some tasks...
- *		  console.log("param:", A(arg));
- *
- *		  return 1234;
- *	   }, function(data) {
- *		  // callback after worker is done
- *	 	  console.log("result:", data);
- *	   },
- *	   [function A(){}]
- *	);
- *
- *	worker(11111);
- * @private
- */
-function runWorker(
-	useWorker = true, fn: Function, callback: Function, depsFn?: Function[]
-): Function {
-	let runFn;
-
-	if (window.Worker && useWorker) {
-		// Web Worker body
-		const blob = new Blob([
-			`${depsFn?.map(String).join(";") ?? ""}
-
-			self.onmessage=function({data}) {
-				const result = (${fn.toString()}).apply(null, data);
-				self.postMessage(result);
-			};`
-		], {
-			type: "text/javascript"
-		});
-		const worker = new Worker(URL.createObjectURL(blob));
-
-		runFn = function(...args) {
-			// trigger worker
-			worker.postMessage(args);
-
-			// listen worker
-			worker.onmessage = function(e) {
-				return callback(e.data);
-			};
-
-			// handle error
-			worker.onerror = function(e) {
-				console.error(e);
-			};
-
-			// return new Promise((resolve, reject) => {
-			// 	worker.onmessage = ({data}) => resolve(data);
-			// 	worker.onerror = reject;
-			// });
-		};
-	} else {
-		runFn = function(...args) {
-			const res = fn(...args);
-
-			callback(res);
-		};
-	}
-
-	return runFn;
 }
 
 /**
