+++ test/internals/boost-spec.ts
@@ -6,6 +6,7 @@
 /* global describe, beforeEach, it, expect */
 import {expect} from "chai";
 import util from "../assets/util";
+import {runWorker} from "../../src/module/worker";
 
 describe("BOOST", () => {
 	let chart;
@@ -66,82 +67,18 @@ describe("BOOST", () => {
 				});
 			});
 		});
-
-
 	});
 
-
-	describe.skip("useWorker", () => {
-		it("for columns data type", done => {
-			const chart = util.generate({
-				boost: {
-					useWorker: true
+	describe("useWorker", () => {
+		it("check if given function run on WebWorker thread.", done => {
+			runWorker(true, function test_for_worker(p) {
+					return `${p}_123`;
 				},
-				data: {
-					columns: [
-						["data1", 300, 350, 300, 0, 0, 0],
-						["data2", 130, 100, 140, 200, 150, 50]
-					],
-					type: "line"
-				},
-				onrendered: function() {
-					expect(this.$.chart.empty()).to.be.false;
+				function(res) {
+					expect(res).to.be.equal("abcd_123");
 					done();
 				}
-			});
-
-			// workers works async, so at this point node references are not yet updated
-			expect(chart.$).to.be.undefined;
-		});
-
-		it("for rows data type", done => {
-			const chart = util.generate({
-				boost: {
-					useWorker: true
-				},
-				data: {
-					rows: [
-						["data1", "data2", "data3"],
-						[90, 120, 300],
-						[40, 160, 240],
-						[50, 200, 290],
-						[120, 160, 230],
-						[80, 130, 300],
-						[90, 220, 320]
-					],
-					type: "line"
-				},
-				onrendered: function() {
-					expect(this.$.chart.empty()).to.be.false;
-					done();
-				}
-			});
-
-			// workers works async, so at this point node references are not yet updated
-			expect(chart.$).to.be.undefined;
-		});
-
-		it("for JSON data type", done => {
-			const chart = util.generate({
-				boost: {
-					useWorker: true
-				},
-				data: {
-					json: {
-						data1: [30, 20, 50, 40, 60, 50],
-						data2: [200, 130, 90, 240, 130, 220],
-						data3: [300, 200, 160, 400, 250, 250]
-					},
-					type: "line"
-				},
-				onrendered: function() {
-					expect(this.$.chart.empty()).to.be.false;
-					done();
-				}
-			});
-
-			// workers works async, so at this point node references are not yet updated
-			expect(chart.$).to.be.undefined;
+			)("abcd");
 		});
 	});
 });
